Updated Service: updated the mount path to be windows compliant
Update Overview:
Mount path from home directory on windows had to be reformated because of the colon and double backslashes. replaced the assignment with a ternary operation that removes windows specific aspects of the home path if user is on windows. Also reformated the trailing forward slash on Mount_path as it was causing poorly formated paths

Update Known Issues:
none

Updates include:
MOUNT_PATH assignment changes