First step toward multi-platform builds.
Another notable change is that the Dockerfile now uses the standard Python slim container rather than the Balena images. For one it's slightly smaller. For another it will automatically pull the correct architecture without needing it in the image name. It should also make it easier to re-enable GPU builds.