Move `pip compile` to Dockerfile.
If any dependencies had variants based on the platform then the wrong requirements would be rendered out. For example, if I used Chassis on a Windows machine and one of the dependencies had a variant that used the `win32` APIs, then the generated `requirements.txt` would include those packages which would not work when they were installed during the Docker build.

Now the dependencies are generated in the same environment that they will be installed in and the `pip compile` and dependency normalization happens in a new pre-stage of the Docker build process.