Make OMIClient async. (#146)
* Make OMIClient async.

This change is made to support running both as a Python script as well as in a Jupyter notebook.

Since Jupyter uses IPython as the runtime, we cannot use the `asyncio` methods to try and run async functions outside of a larger event loop.

This change adds a tiny bit of complexity to using the OMIClient but it should be familiar to anyone who's used async/await in Python.

* Updated docs with new async OMIClient.
# Conflicts:
#	docs/docs/getting-started/index.md